final test passing
tried an ugly looking anonymous class with a handle method override, but IntelliJ suggested a very pretty lambda substitute.  Not sure it's exactly what the Readme meant by "if should reply returns true, the Tracker should have its "handle" method called passing in an instance of EventHandler " -- but as it does the same thing, passes the tests, and is so much nicer looking that the anonymous class method, I'm keeping it this way for now.